+++ src/visits/types/CommonVisitsProps.ts
@@ -1,7 +1,5 @@
-import { SelectedServer } from '../../servers/data';
 import { Settings } from '../../settings/reducers/settings';
 
 export interface CommonVisitsProps {
-  selectedServer: SelectedServer;
   settings: Settings;
 }